hwservicemanager: Make hwservicemanager in cpu foreground group
When camera launch, it has dependency on hwservicemanager, which is
in cpu system group and scheduling latency is not guaranteed. So,
change it to cpu foreground group, which prefers idle cpu and scheduling
latency should improve.

Bug: 187146897
Test: hwservicemanager cpu group changed, 10% improvement in camera
device open time.
Change-Id: I6298b11225c0768c7ff0e9bbbbe6a51e50dfecc5
Merged-In: I6298b11225c0768c7ff0e9bbbbe6a51e50dfecc5